# [Merged by Bors] - use bevy_utils::HashMap for better performance. TypeId is predefined …