@shuoli84 2023-02-12
> LGTM. Could you also try using the FxHash used by Components here too? Could net a bit more of a performance gain.

Ha, I'm just thinking replace the HashMap for component with AHash, so we can drop one dependency.