@shuoli84 2023-02-12
> LGTM. Could you also try using the FxHash used by Components here too? Could net a bit more of a performance gain.

@james7132 I tried FxHash, the performance difference is not obvious. Also fxhash repo is not well maintained compare to ahash. so I modify component's hashmap to ahash, is it ok?