@shuoli84 2023-02-13
@james7132 sure, I'll do a test. The key is figuring out which HashMap is better(runtime perf faster) with Key = u64.  "std::collections::HashMap", "hashbrown::HashMap" combine with Hasher "FxHash", "AHash".